starting to create v1
